Do not require $srctree to be set for non-kernel projects.
(It was never required if you explicitly passed a 'base_dir', but it's a
bit silly to have to do that too.)

This is a bug. I expected os.path.expandvars() to replace references
non-set environment variables with nothing, but it leaves them as is.

Work around it by letting base_dir = None be special and the default. It
uses $srctree if set and the current directory otherwise. This has the
following advantages:

 - It avoids having to reimplement a different version of
   os.path.expandvars() and special-casing "" to mean the current
   directory.

 - It means '$' can appear in paths. (Though it probably never will.)

Maybe the expansion behavior could be removed too, but keep it for now
to be backwards compatible.