Report correct locations in the presence of continuation lines.
The line number was previously for logical lines only. Oversight.

Get rid of _get_lines() and keep the raw lines in _FileFeed instead,
only joining lines with continuation lines as they are fetched. This
makes the index correspond to the correct line number from the file. (It
also means most lines are returned as-is without any logic applied to
them, which is nice.)

Litter tests/Klocation with continuation lines to get test coverage.

Remove some unused functions that were previously inherited from _Feed
and remove it as a base class of _FileFeed.