Massively reduce test suite memory usage
Previously we would load all arches once in the beginning and reuse them
throughout the test suite. With 36 arches with all symbols evaluated and
cached, this used many gigabytes of memory (unless the test suite was
run with pypy). Restructure things so that we only keep one arch loaded
at a time. This adds some reparsing, but it's not a big deal.

Do an ugly piggy-backing of some code cleanup in the test suite (mostly
using un-numbered {} with format()).