Handle choices more like the C implementation
Choice symbols get the choice itself propagated to the conditions of
their properties. This works since the Symbol/Choice value interfaces
are compatible now.

Simplifies value calculations and makes .assignable work automagically
for choice symbols in choices in any mode. One drawback is that a '&&
<choice>' now shows up on conditions when printing choice symbols, which
is not valid Kconfig. That should be easy to work around if needed at
least.