Do not invalidate everything when loading .configs
In replace mode, only unset symbols that didn't get set.

Get rid of _set_value_no_invalidate() and just do a normal set_value()
with invalidation. It's speedy with the new invalidation algorithm, and
simpler.

Not a big performance boost (but a small one), but means that
invalidation must be even more rock solid for the test suite to pass (no
global invalidations to hide behind), which is nice.

Also make y assignments to choice symbol update just the choice user
value. Gives nicer behavior when the choice mode is changed.