Only check for implicit submenus for symbols
Implicit submenus are only ever created with a symbol as the base.
_has_auto_menu_dep() would indirectly reject anything that wasn't a
symbol anyway, but it led to redundant work.

Also get rid of _check_auto_menu() and inline it into _finalize_tree().
It's easier to read without the recursive call being hidden inside
_check_auto_menu().

Tweak the comments a bit too.