Add example that finds references to undefined symbols
Does a global search over all architectures in the kernel, which should
avoid false positives.

Referencing an undefined symbol in a particular architecture can be fine
in a Kconfig file that's shared by multiple architectures, but if the
symbol isn't defined by any architecture, it's likely to be an error (or
a potential cleanup).