Add back omitted $ARCH != $SRCARCH defconfig tests
Oversight during refactoring. defconfig files should be searched for in
in arch/$SRCARCH/ rather than arch/$ARCH/, to account for
$ARCH != $SRCARCH.

Adds a few tests to test_defconfig(). Some of the ARCH/defconfig
combinations might be nonsensical, but that's fine.

Make all test functions take $SRCARCH as well as $ARCH, which makes the
implementation straightforward.