Implement 'rsource' statement ('source' with relative path)
The 'rsource' statement works like 'source', but looks relative to the
Kconfig file that has the 'rsource' rather than relative to the base
Kconfig file. Using 'rsource' makes it possible to move subtrees with
Kconfig files around without breaking references to other Kconfig files.

So far, this is a Kconfiglib-exclusive feature.