Return "" for unwritten symbols in Symbol.config_string
The previous return value was None.

Returning "" makes write_config() neater and feels a bit more Pythonic.
It might simplify the implementation of some planned features as well.

This is a small API break, so the major version will be bumped to 4 in
the next release.

Only code that explicitly tests Symbol.config_string against None will
be affected: 'if sym.config_string is None:' will break, but not
'if not sym.config_string:'.