Properly remember y user values for choice symbols
Commit e8b4ecb ("Don't special-case user_value for choice symbols set to
y") was meant to make user_value reflect an y value assigned to a choice
symbol, but didn't get it right for load_config() with replace=True: The
user value was set at first, but was then unset again due to a misplaced
'_was_set = True' assignment.

This commit makes y choice symbol user values always be remembered. The
correctness of generated files was not affected, since the user
selection of a choice is actually remembered as Choice.user_selection.
The previous commit just turned into a no-op for load_config() with
replace=True.