Decode sources as UTF-8 instead of relying on the system locale
Opening files without specifying an encoding will cause the system
locale to determine the encoding. It is undesired behaviour IMHO that
reading/decoding Kconfig sources is platform-dependent, instead
kconfiglib should behave in the same way on all platforms.

Example:

$ export LANG=C
$ echo "# Author: Sebastian Bøe" >> Kconfig
$ test_kconfig_lib.py

Alternatively, one could decode as ascii, especially if it is
explicitly documented that Kconfig sources (including comments) must
be ascii.

The most important thing is that kconfiglib behaves the same on all
platforms.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>