Add a terminal menuconfig implementation
Uses the 'curses' module, which is in the Python standard library. Only
Python 3 is supported, mostly due to Unicode support.

Windows support is provided through an external package which adds
support for the 'curses' module on Windows.

See the changes and the docstring in the newly added menuconfig.py for
more information.

Piggyback a README notes mentioning that Kconfiglib supports Unicode.