Make warnings available in a list
Also make printing of warnings to stderr optional. It's on by default,
since it's almost always what you want.

This makes printing and processing of warnings more flexible, compared
to always printing warnings to stderr as soon as they're generated. It
was inspired by wanting to promote certain warnings to errors in Zephyr,
which previously required capturing stderr.