Force encoding to UTF-8 by default on Python 3
Enough people have run into exceptions due to running in the C locale
that it seems worthwhile.

Add a new 'encoding' parameter to Kconfig.__init__() that specifies the
encoding to use and make it default to "utf-8". Passing None gives the
old behavior of using the encoding specified in the environment.

Related PEP: https://www.python.org/dev/peps/pep-0538/