menuconfig: Add a mode that shows invisible items
When show-all mode is enabled, all items in the current menu are shown,
including promptless and invisible items. Promptless items are shown
with their name within <> brackets in place of the prompt.

This will make it possible to jump to invisible items once the jump-to
feature is added. It is also a handy feature on its own.

.config loading from within the interface will make use of it as well,
in case the current menu becomes complete invisible.

Piggyback various code cleanups. Rename _visible* to _shown*, as it now
includes invisible items in show-all mode.