menuconfig: Convert the C locale to a UTF-8 locale for LC_CTYPE
If LC_CTYPE is set to the C locale, try to convert it to a UTF-8 locale.
Use a list of commonly available UTF-8 locales. Give up and use the C
locale if none of them are available.

This fixes curses Unicode I/O issues on systems with bad defaults:
ncurses configures itself from the locale settings, and the C locale
implies ASCII.

The logic mirrors https://www.python.org/dev/peps/pep-0538/. I took the
list of locales to try from the CPython code (in Python/pylifecycle.c).