Restore compatibility with $UNAME_RELEASE
UNAME_RELEASE is expanded in one of the 'default's of the DEFCONFIG_LIST
symbol in the Linux kernel. This broke after "$FOO" was changed to
directly expand to the value of the environment variable FOO, rather
than to the value of the symbol FOO.

Restore compatibility with a small wrapper.