Simplify allyesconfig
Setting all tristate choice symbols to 'm' is enough to always do the
right thing in practice. It has no effect for bool choice symbols (since
'm' will be ignored), but forces symbols in m-mode choices to 'm'.

'm' will also be ignored for tristate choice symbols in choices in 'y'
mode, because tristate is automagically converted to bool there.