Look up the top-level Kconfig file relative to $srctree
Due to an old design braino, the top-level Kconfig filename passed to
Kconfig.__init__() wasn't looked up relative to $srctree, breaking
out-of-tree usage for e.g. menuconfig. Fixing it required ugliness like

  srctree = os.environ.get("srctree", "")
  kconfiglib.Kconfig(os.path.join(srctree, "Kconfig"))

Change the behavior of Kconfig.__init__() to look up the top-level
Kconfig file relative to $srctree. This means that all Kconfig files
(both the top-level file and any source'd files) now use $srctree, which
makes the vast majority of scripts just work when running out-of-tree.

Also remove the note re. loading a subset of Kconfig files. Saying that
the top-level file and all source'd Kconfig files are looked up relative
to $srctree should make the behavior clear enough.

Print a note about the new behavior whenever the top-level Kconfig file
can't be opened, as this change could be breaking for some scripts.

This is a slight backwards-compatiblity break, so the major version will
be bumped.