Revert "Fix $srctree logic for the top-level Kconfig file"
This reverts commit 8a3999bc708e8468ff79665e3cbdfccd603160e1.

I realized that this should go in a major release at least, because it
has the potential to break scripts that rely on the old behavior of
ignoring $srctree for the filename passed to Kconfig.__init__().

A dummy release will bump the version to 9.4.2. Any future release with
the change will be 10.0.0.