Make 10.0.0 release
Adds commit 8a3999b ("Fix $srctree logic for the top-level Kconfig
file"), which fixes an old braino: The top-level Kconfig file passed to
Kconfig.__init__() wasn't looked up relative to $srctree, breaking
out-of-tree operation for scripts unless ugly manual $srctree hacks were
added.

All Kconfig files are now consistently looked up relative to $srctree.

This is a backwards-incompatible change (though with a simple fix), so
the major version is bumped. A note is printed on errors that might be
due to the changed behavior.