Add a generic node iterator.
Use the iterator to traverse symbols/nodes in write_config(),
write_min_config(), write_autoconfs()...

This ensures that the order of symbols is consistent between config
files and headers and that symbols don't get shuffled around between
configuration changes.

A nice side effect is that comments are now included in the generated
header files and minimal configs.