menuconfig: Don't return when selecting a choice symbol with children
If a symbol depends on a choice symbol that immediately precedes it,
it's put into an implicit submenu and not considered a choice symbol.
Don't immediately return to the parent menu when selecting such a
symbol. It makes the children hard to spot.

This is used (or was... not sure if it still is) in a single Kconfig
file in the kernel. It's completely undocumented.