menuconfig: Do not use red for invisible symbols outside of show-all mode
Invisible symbols can show up outside show-all mode if an invisible
symbol has visible symbols in an implicit (indented) submenu. Printing
them in red like in show-all mode is the perfectly consistent thing to
do, but might look confusing/broken.

Only print invisible symbols in red in show-all mode.

Also change the bracket style for all invisible bool/tristate symbols to

  - - N_SYM
  -M- M_SYM
  -*- Y_SYM

That makes it clear that they can't be changed, even when they show up
outside show-all mode, and makes logical sense if that style is taken to
always mean that the symbol value can't be changed.

You lose a quick reference for whether a symbol is a bool or tristate,
but it's not a huge deal.