menuconfig: Print warnings instead of having StyleError
Mostly motivated by curses.COLORS, which depends on the terminal.

It's probably friendlier to run even if the style gets a bit messed up,
but it's also nice to point out problems.

Missing style templates and missing styles to the left of '=' are
pointed out now too, with the help of some flag hackery.

Prefix the warnings with 'menuconfig warning:' to make it easier to
distinguish them from warnings from Kconfiglib.