Fix parse error message on the line after help texts
The correct error and line number was reported, but not the correct line
contents. self._line needs to be set before calling _tokenize(), so that
_parse_error() knows the context.

There's no need to set self._line for empty lines, because we
immediately end up back in _next_line() after parsing the help text,
which refetches the empty line and updates self._line.