genconfig: Allow writing a full .config without incremental build info
For projects that don't use incremental build information, it's a bit
wasteful to generate it just to get an updated .config file to include
in Makefiles. Add a '--config-out <filename>' option that just writes
the updated .config file, for those cases.

Also remove a redundant default=None argument for --sync-deps. None is
the default value.