menuconfig: Improve/fix promptless choice handling
The code assumed that all parent (interface) menus always have a prompt,
which is false for items in promptless choices. This lead to a crash
e.g. when viewing the symbol information for a symbol within a
promptless choice.

Promptless choices with children can show up "legitimately" when people
define choices in multiple locations to add symbols, though this is
broken in the C tools.

Use standard_sc_expr_str(node.item) instead of the non-existing prompt
for promptless choices. That way they show up as
<choice (name if any>)>, which is consistent with how they're shown
elsewhere.

This commit also changes how choices names are displayed in
show-name/show-all mode, to the standard_sc_expr_str() format.