menuconfig: Add jump-to for choices, menus, and comments
For choices, search the name and the prompt. This is the same as for
symbols, except names are optional (and rare) for choices.

For menus and comments, search the prompt (title text).

When jumping to a non-empty choice or menu, jump into it instead of
jumping to its menu node. If show-all mode is off and there are visible
items in the choice/menu, then jump to the first visible node.
Otherwise, enable show-all and jump to the first node.