menuconfig: Fix some minor resizing jumpiness on gnome-terminal
gnome-terminal jerks a bit whenever the terminal is shrunk so that the
row with the cursor on it disappears. This is probably gnome-terminal
shoddiness, but it's pretty easy to work around by drawing stuff at the
top last, so do it.

No need to do it for the jump-to dialog, because the cursor is already
in the edit box at the top there.