menuconfig: Support HOME and END in the jump-to dialog
Oversight. END can be handy for viewing all choices, menus, and
comments, which appear at the end.

Also support Ctrl-D as an alias for Page Down. Ctrl-U is already used by
the edit box (erase to beginning of line).

Also change the jump-to dialog title to hint that other things besides
symbols can be jumped to.