Add support for KCONFIG_ALLCONFIG
This allows some symbol values to be forced while running
all{def,no,yes,mod}config.py. See Documentation/kbuild/kconfig.txt in
the Linux kernel.

Add a helper function load_allconfig() to Kconfiglib to avoid code
duplication in the tools.

Also add functions for enabling/disabling the warning that's generated
when a symbol is assigned multiple times in a (set of) .config files and
the values differ. It should be disabled when merging the
KCONFIG_ALLCONFIG configuration file. Previously, only the warning
generated when the assigned values are identical could be disabled.

Disable all warnings related to assigning a symbol multiple times in
examples/merge_config.py as well.