menuconfig: Never snap scroll back to max_scroll
When scroll > max_scroll (blank space at the bottom of the menu, and
also scrolled down), moving the cursor down would snap the scroll back
to max_scroll, with a small annoying jump. This could happen when going
into show-all mode and moving the cursor down near the bottom of a menu,
for example.

Fix it by leaving the scroll as-is when scroll >= max_scroll (increasing
it by one otherwise).