Use 'foo.__class__ is Bar' instead of 'isinstance(foo, Bar)'
This is 30%-60% faster for both the matching and non-matching case, as
measured with timeit on Python 2.7 and 3.6, and saves at least a few
percent of total parsing time (and probably some evaluation time too).

isinstance(foo, tuple) is particularly slow.

Symbol and Choice instances are always created by us, so potential
subclassing shouldn't be a problem.