Use constants when _tokens_i is known
Same deal as for the initial token, except we sometimes know that we're
dealing with the second token as well.

Inline _expect_nonconst_sym_and_eol() and _expect_str(), which are
single-use. That allows more specific error messages to be used as well.

Also tweak an outdated comment in _tokenize() re. None-termination.
Token fetching is more manual now.